refactor(Consensus): Factor CUP functionality out of dkg submodule (#3126)
The functions `make_registry_cup` and `make_registry_cup_from_cup` are
used by other crates to create registry CUPs.

This PR moves the functions out of the `dkg` submodule into its own
module `cup_utils` and directly exports them from the top-level
`ic_consensus`.

The reason is that these functions also use `idkg` functionality.

---------
