+++ rs/consensus/src/lib.rs
@@ -8,5 +8,8 @@
 mod bouncer_metrics;
 pub mod certification;
 pub mod consensus;
+pub mod cup_utils;
 pub mod dkg;
 pub mod idkg;
+
+pub use cup_utils::{make_registry_cup, make_registry_cup_from_cup_contents};
